refactor(agentic-ai): fold Anthropic client construction into AnthropicChatModelApiFactory
Drop the AnthropicOkHttpClientFactory class and move its buildClient logic
directly into AnthropicChatModelApiFactory as a package-private static method,
since it had exactly one caller and its own construction wasn't a meaningful
seam. The dedicated wire-level test moves with it (renamed to
AnthropicChatModelApiFactoryClientTest), still calling the method directly.

Stop sending a placeholder x-api-key for compatible backends configured with
no authentication: verified empirically that the SDK client builder neither
requires an api key to build nor to send a request without one, so the
placeholder was unnecessary; the header is now simply omitted.

Clarify why compatible-backend headers/queryParameters are applied at the
client level while requestParameters are applied per-request in the message
converter.
