+++ connectors/agentic-ai/connector-agentic-ai/src/test/java/io/camunda/connector/agenticai/aiagent/chatmodel/provider/anthropic/AnthropicMessageRequestConverterTest.java
@@ -425,9 +425,9 @@ class AnthropicMessageRequestConverterTest {
       "deprecation") // temperature()/topP()/topK() deprecated in anthropic-java 2.48.0
   void mapsFullV1ParameterParitySet() {
     // v1 Anthropic exposed exactly: endpoint, apiKey, timeout, model, maxTokens, temperature,
-    // topP, topK. endpoint/apiKey/timeout are transport-layer (AnthropicOkHttpClientFactory,
-    // Task 2's concern); this asserts the remaining 5 model-parameter fields this converter is
-    // responsible for.
+    // topP, topK. endpoint/apiKey/timeout are transport-layer (AnthropicChatModelApiFactory's
+    // concern); this asserts the remaining 5 model-parameter fields this converter is responsible
+    // for.
     final var parameters = new AnthropicModelParameters(2048, 0.5, 0.9, 40, null, null);
     final var snapshot = new ConversationSnapshot(List.of(), List.of());
